+++ iot_firmware/event/enum.py
@@ -1,4 +1,7 @@
 """Event Enums."""
+from enum import auto
+from enum import Enum
+
 from iot_firmware.enums import StrEnum
 
 
@@ -10,3 +13,11 @@ class EventLevel(StrEnum):
     WARNING = "WARNING"
     ERROR = "ERROR"
     CRITICAL = "CRITICAL"
+
+
+class EventHandlerState(Enum):
+    """EventHandler possible states."""
+
+    IDLE = auto()
+    RUNNING = auto()
+    STOPPING = auto()