+++ tests/test_event.py
@@ -1,3 +1,4 @@
+import asyncio
 import uuid
 from typing import Any
 
@@ -16,14 +17,14 @@ from mocks.mocks import MockObject
 def test_event_handler_subscribe_simple_function():
     event_handler = EventHandler()
     event_handler.subscribe(MockEvent, mock_function)
-    assert MockEvent.type.uuid in event_handler.subscribers
+    assert MockEvent.type.uuid in event_handler._subscribers
 
 
 def test_event_handler_subscribe_class_function():
     event_handler = EventHandler()
     mock_object = MockObject()
     event_handler.subscribe(MockEvent, mock_object.mock_function)
-    assert MockEvent.type.uuid in event_handler.subscribers
+    assert MockEvent.type.uuid in event_handler._subscribers
 
 
 def test_event_handler_subscribe_bad_class():
@@ -87,14 +88,14 @@ def test_event_handler_unsubscribe_add_bad_class_fn():
     with pytest.raises(TypeError) as exc_info:
         event_handler.unsubscribe(CustomEvent, None)
 
-    assert len(exc_info.value.args) == 2
+    assert exc_info.value.args
 
 
 def test_event_handler_unsubscribe():
     event_handler = EventHandler()
     event_handler.subscribe(MockEvent, mock_function)
     event_handler.unsubscribe(MockEvent, mock_function)
-    assert MockEvent.type.uuid not in event_handler.subscribers
+    assert MockEvent.type.uuid not in event_handler._subscribers
 
 
 def test_event_handler_publish_with_subscribers():
@@ -110,8 +111,19 @@ def test_event_handler_publish_no_event():
 
     mock_event = CustomEvent()
     event_handler = EventHandler()
-    with pytest.raises(TypeError):
-        event_handler.publish(mock_event)
+    event_handler.publish(mock_event)
+    assert event_handler._event_buffer.qsize() == 0
+
+
+def test_event_handler_publish_no_event_with_subscriber():
+    class CustomEvent:
+        pass
+
+    mock_event = CustomEvent()
+    event_handler = EventHandler()
+    event_handler.subscribe(MockEvent, mock_function)
+    event_handler.publish(mock_event)
+    assert event_handler._event_buffer.qsize() == 0
 
 
 def test_event_handler_publish_without_subscribers():
@@ -184,3 +196,172 @@ def test_event_class():
     bar = BarEvent()
 
     assert foo.uuid != bar.uuid
+
+
+def test_unsubscribe_a_not_subscribed_event():
+    event_handler = EventHandler()
+    event_handler.unsubscribe(MockEvent, mock_function)
+
+
+def test_event_handler_is_full():
+    event_handler = EventHandler(buffer_maxsize=2)
+    assert not event_handler._event_buffer.full()
+
+    event_handler.subscribe(MockEvent, mock_function)
+    event_handler.publish(MockEvent())
+    assert not event_handler._event_buffer.full()
+
+    event_handler.publish(MockEvent())
+    assert event_handler._event_buffer.full()
+
+    event_handler.publish(MockEvent())
+    assert event_handler._event_buffer.full()
+
+
+@pytest.mark.asyncio
+async def test_event_handler_worker_workflow_function():
+    event_handler = EventHandler(buffer_maxsize=2, num_workers=1)
+    event_handler.subscribe(MockEvent, mock_function)
+
+    event = MockEvent(data=0)
+    event_handler.publish(event)
+
+    try:
+        await asyncio.wait_for(event_handler.run(), timeout=0.3)
+    except asyncio.TimeoutError:
+        pass
+
+    assert event.data == 1
+
+
+@pytest.mark.asyncio
+async def test_event_handler_worker_timeout():
+    event_handler = EventHandler(
+        buffer_maxsize=2, num_workers=1, worker_timeout_seconds=0.5
+    )
+
+    async def quick_function(event: MockEvent):
+        await asyncio.sleep(0.1)
+        event.data += 1
+
+    async def slow_function(event: MockEvent):
+        await asyncio.sleep(1)
+        event.data += 1
+
+    event_handler.subscribe(MockEvent, quick_function)
+    event_handler.subscribe(MockEvent, slow_function)
+
+    mock_event = MockEvent(data=0)
+    event_handler.publish(mock_event)
+
+    try:
+        await asyncio.wait_for(event_handler.run(), timeout=1)
+    except asyncio.TimeoutError:
+        pass
+
+    assert mock_event.data == 1
+
+
+@pytest.mark.asyncio
+async def test_event_handler_stop_workers():
+    event_handler = EventHandler(num_workers=2)
+
+    async def normal_function(event: MockEvent):
+        await asyncio.sleep(0.5)
+        event.data += 1
+
+    event_handler.subscribe(MockEvent, normal_function)
+
+    mock_event = MockEvent(data=0)
+    event_handler.publish(mock_event)
+
+    async def stop_event_handler(event_handler_local: EventHandler):
+        await asyncio.sleep(0.2)
+        await event_handler_local.stop()
+
+    await asyncio.gather(event_handler.run(), stop_event_handler(event_handler))
+
+    assert mock_event.data == 1
+    assert len(event_handler._workers) == 0
+
+
+@pytest.mark.asyncio
+async def test_event_handler_stopping_workers():
+    event_handler = EventHandler(num_workers=1, buffer_maxsize=1)
+
+    async def normal_function(event: MockEvent):
+        await asyncio.sleep(0.5)
+        event.data += 1
+
+    event_handler.subscribe(MockEvent, normal_function)
+
+    mock_event = MockEvent(data=0)
+    event_handler.publish(mock_event)
+
+    async def stop_event_handler(event_handler_local: EventHandler):
+        await asyncio.sleep(0.2)
+        await event_handler_local.stop()
+        event_handler.publish(mock_event)
+        await asyncio.sleep(1)
+        event_handler.publish(mock_event)
+
+    await asyncio.gather(event_handler.run(), stop_event_handler(event_handler))
+
+    assert mock_event.data == 1
+    assert len(event_handler._workers) == 0
+
+
+@pytest.mark.asyncio
+async def test_event_handler_stop_workers_error():
+    event_handler = EventHandler()
+
+    with pytest.raises(RuntimeError):
+        await event_handler.stop()
+
+
+@pytest.mark.asyncio
+async def test_event_handler_cancel_workers():
+    event_handler = EventHandler(num_workers=2)
+
+    async def normal_function(event: MockEvent):
+        await asyncio.sleep(0.5)
+        event.data += 1
+
+    event_handler.subscribe(MockEvent, normal_function)
+
+    mock_event = MockEvent(data=0)
+    event_handler.publish(mock_event)
+
+    async def cancel_event_handler(event_handler_local: EventHandler):
+        await asyncio.sleep(0.2)
+        event_handler_local.cancel()
+
+    await asyncio.gather(event_handler.run(), cancel_event_handler(event_handler))
+
+    assert mock_event.data == 0
+    assert len(event_handler._workers) == 0
+
+
+def test_event_handler_cancel_workers_error():
+    event_handler = EventHandler()
+
+    with pytest.raises(RuntimeError):
+        event_handler.cancel()
+
+
+@pytest.mark.asyncio
+async def test_event_handler_raise_error_in_function():
+    event_handler = EventHandler(num_workers=2)
+
+    async def function_raise_error(event: MockEvent):
+        raise RuntimeError("error triggered")
+
+    event_handler.subscribe(MockEvent, function_raise_error)
+
+    mock_event = MockEvent(data=0)
+    event_handler.publish(mock_event)
+
+    try:
+        await asyncio.wait_for(event_handler.run(), timeout=0.5)
+    except asyncio.TimeoutError:
+        pass