+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
   - Python package (iot_firmware)
     - `setup.cfg` and `pyproject.toml`
     - Communications API schema
-    - Sync Event Handler
+    - Async Event Handler
     - Docstrings
   - Sphinx Documentation (docs) (with diagrams)
   - Tests (pytest)